moneymong-147 fix: AgencyInviteCodeView state, viewModel 수정

+++ data/src/main/java/com/moneymong/moneymong/data/datasource/AgencyRemoteDataSource.kt
@@ -13,6 +13,6 @@ interface AgencyRemoteDataSource {
 
     suspend fun getAgencies(page: Int, size: Int): Result<AgenciesGetResponse>
 
-    suspend fun agencyCodeNumbers(agencyId: Int, codeNumbers: AgencyJoinRequest) : Result<AgencyJoinResponse>
+    suspend fun agencyCodeNumbers(agencyId: Long, codeNumbers: AgencyJoinRequest) : Result<AgencyJoinResponse>
 
 }
